+++ types/index.d.ts
@@ -84,10 +84,10 @@ export class XboxSession extends EventEmitter<{
   invite(identifier: UserIdentifier, options?: OperationOptions): Promise<void>
   close(): Promise<void>
 }
-export class RtaSubscription<T = unknown> extends EventEmitter {
+export class XboxRTASubscription<T = unknown> extends EventEmitter {
   private constructor()
   readonly uri: string
-  readonly data: T
+  readonly initialData: T
   readonly closed: boolean
   close(): Promise<void>
   on(event: 'ready' | 'data', listener: (data: T) => void): this
@@ -114,10 +114,11 @@ export class XboxRTASocket extends EventEmitter {
   constructor(authflow: XboxTokenProvider)
   connect(options?: OperationOptions): Promise<void>
   reconnect(): Promise<void>
-  subscribe<T = unknown>(uri: string, options?: OperationOptions): Promise<RtaSubscription<T>>
+  subscribe<T = unknown>(uri: string, options?: OperationOptions): Promise<XboxRTASubscription<T>>
   close(): Promise<void>
   on(event: 'resync', listener: () => void): this
-  on(event: 'close', listener: (code: number, reason: string) => void): this
+  on(event: 'disconnect', listener: (code: number, reason: string) => void): this
+  on(event: 'close', listener: () => void): this
   on(event: 'error', listener: (error: Error) => void): this
   on(event: string | symbol, listener: (...args: any[]) => void): this
 }